refactor(portal): 创建战报时带入当前团UUID
