+++ src/portal/model/trpg.ts
@@ -1,31 +1,8 @@
 import { request } from '@portal/utils/request';
-import { ChatLogItem } from './chat';
 import { hpack, hunpack } from '@shared/utils/json-helper';
 import _isNil from 'lodash/isNil';
 import _isEmpty from 'lodash/isEmpty';
-
-// 消息log必须的字段
-// 用于消息数据压缩
-export const reportLogRequireKey = [
-  'uuid',
-  'sender_uuid',
-  'message',
-  'type',
-  'date',
-  'data',
-  'revoke',
-] as const;
-
-export type ReportLogItem = Pick<
-  ChatLogItem,
-  typeof reportLogRequireKey[number]
->;
-export interface EditLogItem extends ReportLogItem {
-  selected: boolean;
-}
-export interface DetailLogItem extends ReportLogItem {
-  isShow: boolean;
-}
+import { ReportLogItem } from '@shared/model/trpg';
 
 export interface GameReport {
   uuid: string;
@@ -36,25 +13,6 @@ export interface GameReport {
     logs: ReportLogItem[];
   };
 }
-
-/**
- * 创建游戏战报
- * @param title 标题
- * @param playerUUID 战报主视角的UUID
- * @param logs 记录
- */
-export const createTRPGGameReport = async (
-  title: string,
-  playerUUID: string,
-  logs: ReportLogItem[]
-): Promise<string> => {
-  const { data } = await request.post('/trpg/game-report/create', {
-    title,
-    content: { playerUUID, logs: hpack(logs) },
-  });
-
-  return data.uuid;
-};
 
 /**
  * 获取团战报
