+++ src/shared/model/chat.ts
